+++ pyanaconda/bootloader/yaboot.py
@@ -1,0 +1,183 @@
+#
+# Copyright (C) 2019 Red Hat, Inc.
+#
+# This copyrighted material is made available to anyone wishing to use,
+# modify, copy, or redistribute it subject to the terms and conditions of
+# the GNU General Public License v.2, or (at your option) any later version.
+# This program is distributed in the hope that it will be useful, but WITHOUT
+# ANY WARRANTY expressed or implied, including the implied warranties of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU General
+# Public License for more details.  You should have received a copy of the
+# GNU General Public License along with this program; if not, write to the
+# Free Software Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA
+# 02110-1301, USA.  Any Red Hat trademarks that are incorporated in the
+# source code or documentation are not subject to the GNU General Public
+# License and may only be used or replicated with the express permission of
+# Red Hat, Inc.
+#
+import os
+
+from blivet.devicelibs import raid
+
+from pyanaconda.bootloader.base import BootLoaderError, Arguments, BootLoader
+from pyanaconda.bootloader.image import LinuxBootLoaderImage
+from pyanaconda.core import util
+from pyanaconda.product import productName
+
+from pyanaconda.anaconda_loggers import get_module_logger
+log = get_module_logger(__name__)
+
+__all__ = ["MacYaboot"]
+
+
+# Inherit abstract methods from BootLoader
+# pylint: disable=abstract-method
+class YabootBase(BootLoader):
+    """A base class for yaboot-based boot loaders."""
+
+    def write_config_password(self, config):
+        if self.password:
+            config.write("password=%s\n" % self.password)
+            config.write("restricted\n")
+
+    def write_config_images(self, config):
+        for image in self.images:
+            if not isinstance(image, LinuxBootLoaderImage):
+                # mac os images are handled specially in the header on mac
+                continue
+
+            args = Arguments()
+            if self.password:
+                args.add("rd.shell=0")
+            if image.initrd:
+                initrd_line = "\tinitrd=%s/%s\n" % (self.boot_prefix,
+                                                    image.initrd)
+            else:
+                initrd_line = ""
+
+            root_device_spec = image.device.fstab_spec
+            if root_device_spec.startswith("/"):
+                root_line = "\troot=%s\n" % root_device_spec
+            else:
+                args.add("root=%s" % root_device_spec)
+                root_line = ""
+
+            args.update(self.boot_args)
+            log.info("bootloader.py: used boot args: %s ", args)
+
+            stanza = ("image=%(boot_prefix)s%(kernel)s\n"
+                      "\tlabel=%(label)s\n"
+                      "\tread-only\n"
+                      "%(initrd_line)s"
+                      "%(root_line)s"
+                      "\tappend=\"%(args)s\"\n\n"
+                      % {"kernel": image.kernel, "initrd_line": initrd_line,
+                         "label": self.image_label(image),
+                         "root_line": root_line, "args": args,
+                         "boot_prefix": self.boot_prefix})
+            config.write(stanza)
+
+
+class Yaboot(YabootBase):
+    name = "Yaboot"
+    _config_file = "yaboot.conf"
+    prog = "ybin"
+    image_label_attr = "short_label"
+    packages = ["yaboot"]
+
+    # stage2 device requirements
+    stage2_device_types = ["partition", "mdarray"]
+    stage2_raid_levels = [raid.RAID1]
+
+    #
+    # configuration
+    #
+
+    @property
+    def config_dir(self):
+        conf_dir = "/etc"
+        if self.stage2_device.format.mountpoint == "/boot":
+            conf_dir = "/boot/etc"
+        return conf_dir
+
+    @property
+    def config_file(self):
+        return "%s/%s" % (self.config_dir, self._config_file)
+
+    def write_config_header(self, config):
+        if self.stage2_device.type == "mdarray":
+            boot_part_num = self.stage2_device.parents[0].parted_partition.number
+        else:
+            boot_part_num = self.stage2_device.parted_partition.number
+
+        # yaboot.conf timeout is in tenths of a second. Brilliant.
+        header = ("# yaboot.conf generated by anaconda\n\n"
+                  "boot=%(stage1dev)s\n"
+                  "init-message=\"Welcome to %(product)s!\\nHit <TAB> for "
+                  "boot options\"\n\n"
+                  "partition=%(part_num)d\n"
+                  "timeout=%(timeout)d\n"
+                  "install=/usr/lib/yaboot/yaboot\n"
+                  "delay=5\n"
+                  "enablecdboot\n"
+                  "enableofboot\n"
+                  "enablenetboot\n"
+                  % {"stage1dev": self.stage1_device.path,
+                     "product": productName, "part_num": boot_part_num,
+                     "timeout": self.timeout * 10})
+        config.write(header)
+        self.write_config_variant_header(config)
+        self.write_config_password(config)
+        config.write("\n")
+
+    def write_config_variant_header(self, config):
+        config.write("nonvram\n")
+        config.write("mntpoint=/boot/yaboot\n")
+        config.write("usemount\n")
+
+    def write_config_post(self):
+        super().write_config_post()
+
+        # make symlink in /etc to yaboot.conf if config is in /boot/etc
+        etc_yaboot_conf = util.getSysroot() + "/etc/yaboot.conf"
+        if not os.access(etc_yaboot_conf, os.R_OK):
+            try:
+                os.symlink("../boot/etc/yaboot.conf", etc_yaboot_conf)
+            except OSError as e:
+                log.error("failed to create /etc/yaboot.conf symlink: %s", e)
+
+    def write_config(self):
+        if not os.path.isdir(util.getSysroot() + self.config_dir):
+            os.mkdir(util.getSysroot() + self.config_dir)
+
+        # this writes the config
+        super().write_config()
+
+    #
+    # installation
+    #
+
+    def install(self, args=None):
+        args = ["-f", "-C", self.config_file]
+        rc = util.execInSysroot(self.prog, args)
+        if rc:
+            raise BootLoaderError("boot loader installation failed")
+
+
+class MacYaboot(Yaboot):
+    prog = "mkofboot"
+    can_dual_boot = True
+
+    #
+    # configuration
+    #
+
+    def write_config_variant_header(self, config):
+        try:
+            mac_os = [i for i in self.chain_images if i.label][0]
+        except IndexError:
+            pass
+        else:
+            config.write("macosx=%s\n" % mac_os.device.path)
+
+        config.write("magicboot=/usr/lib/yaboot/ofboot\n")
